# api/docs/openapi.yaml
openapi: 3.0.0
info:
    title: 'Zyvo API'
    version: 1.0.0
    description: "\n        A secure Node.js API server built with Express.js that implements \n        security best practices including CORS, CSRF protection, rate limiting, \n        and comprehensive error handling.\n        \n        ## Features\n        - JWT Authentication\n        - Role-based Access Control\n        - Rate Limiting\n        - CSRF Protection\n        - Input Validation\n        - Comprehensive Error Handling\n        \n        ## Security\n        This API implements multiple layers of security:\n        - CORS with configurable origins\n        - CSRF token validation\n        - Rate limiting per IP\n        - Secure HTTP headers via Helmet.js\n        - Input sanitization and validation\n      "
    contact:
        name: 'Zyvo API Support'
        email: support@zyvo.com
        url: 'https://zyvo.com'
    license:
        name: ISC
        url: 'https://opensource.org/licenses/ISC'
servers:
    -
        url: 'http://localhost:3000'
        description: 'Development server'
    -
        url: 'https://api.zyvo.com'
        description: 'Production server'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'JWT Bearer token authentication'
        csrfToken:
            type: apiKey
            in: header
            name: X-CSRF-Token
            description: 'CSRF token for protection against CSRF attacks'
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        Option:
            type: object
            required: [option_type, option_value]
            properties: {_id: {type: string, description: 'The auto-generated id of the option'}, option_type: {type: string, description: 'The type of option (e.g., Color, Size, Weight)', maxLength: 50}, option_value: {type: string, description: 'The value of the option (e.g., Red, Large, 500g)', maxLength: 100}, name: {type: string, description: 'Display name for the option (defaults to option_value)', maxLength: 100}, slug: {type: string, description: 'Auto-generated URL-friendly identifier'}, is_active: {type: boolean, description: 'Whether the option is currently active', default: true}, sort_order: {type: number, description: 'Sort order for displaying options', default: 0}, full_name: {type: string, description: 'Virtual field combining option_type and name'}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}}
            example: {_id: 60d0fe4f5311236168a109ca, option_type: Color, option_value: Red, name: 'Bright Red', slug: color-red, is_active: true, sort_order: 1, full_name: 'Color: Bright Red', createdAt: '2023-06-22T09:30:00.000Z', updatedAt: '2023-06-22T09:30:00.000Z'}
        OptionType:
            type: object
            properties: {option_type: {type: string, description: 'The type of option'}, values: {type: array, items: {type: object, properties: {_id: {type: string}, option_value: {type: string}, name: {type: string}, slug: {type: string}, sort_order: {type: number}}}}}
            example: {option_type: Color, values: [{_id: 60d0fe4f5311236168a109ca, option_value: Red, name: 'Bright Red', slug: color-red, sort_order: 1}]}
        OptionResponse:
            type: object
            properties: {success: {type: boolean}, message: {type: string}, data: {$ref: '#/components/schemas/Option'}}
        OptionsListResponse:
            type: object
            properties: {success: {type: boolean}, data: {type: array, items: {$ref: '#/components/schemas/Option'}}, pagination: {type: object, properties: {currentPage: {type: number}, totalPages: {type: number}, totalItems: {type: number}, itemsPerPage: {type: number}, hasNextPage: {type: boolean}, hasPrevPage: {type: boolean}}}}
...
paths:
...
    /api/v1/options:
        post:
            summary: 'Create a new option'
            tags: [Options]
            security: [{BearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [option_type, option_value], properties: {option_type: {type: string, maxLength: 50, example: Color}, option_value: {type: string, maxLength: 100, example: Red}, name: {type: string, maxLength: 100, example: 'Bright Red'}, sort_order: {type: number, minimum: 0, example: 1}}}}}}
            responses: {'201': {description: 'Option created successfully', content: {application/json: {schema: {$ref: '#/components/schemas/OptionResponse'}}}}, '400': {description: 'Validation error or duplicate option'}, '401': {description: 'Unauthorized - Admin access required'}, '500': {description: 'Internal server error'}}
        get:
            summary: 'Get all options with filtering and pagination'
            tags: [Options]
            parameters: [{in: query, name: page, schema: {type: integer, minimum: 1}, description: 'Page number for pagination'}, {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100}, description: 'Number of items per page'}, {in: query, name: option_type, schema: {type: string, maxLength: 50}, description: 'Filter by option type (case-insensitive)'}, {in: query, name: is_active, schema: {type: boolean}, description: 'Filter by active status (admin only)'}, {in: query, name: search, schema: {type: string, maxLength: 100}, description: 'Search in name, option_value, and option_type'}, {in: query, name: sort, schema: {type: string, enum: [option_type, name, sort_order]}, description: 'Sort field'}, {in: query, name: order, schema: {type: string, enum: [asc, desc]}, description: 'Sort order'}, {in: query, name: include_inactive, schema: {type: boolean}, description: 'Include inactive options (admin only)'}]
            responses: {'200': {description: 'List of options retrieved successfully', content: {application/json: {schema: {$ref: '#/components/schemas/OptionsListResponse'}}}}, '400': {description: 'Invalid query parameters'}, '500': {description: 'Internal server error'}}
    /api/v1/options/types:
        get:
            summary: 'Get all option types with their values'
            tags: [Options]
            parameters: [{in: query, name: include_inactive, schema: {type: boolean}, description: 'Include inactive options (admin only)'}]
            responses: {'200': {description: 'Option types retrieved successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, data: {type: array, items: {$ref: '#/components/schemas/OptionType'}}}}}}}, '500': {description: 'Internal server error'}}
    /api/v1/options/stats:
        get:
            summary: 'Get option statistics'
            tags: [Options]
            security: [{BearerAuth: []}]
            responses: {'200': {description: 'Option statistics retrieved successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, data: {type: object, properties: {totalOptions: {type: number}, activeOptions: {type: number}, inactiveOptions: {type: number}, totalOptionTypes: {type: number}, topOptionTypes: {type: array, items: {type: object, properties: {option_type: {type: string}, count: {type: number}}}}}}}}}}}, '401': {description: 'Unauthorized - Admin access required'}, '500': {description: 'Internal server error'}}
    '/api/v1/options/{id}':
        get:
            summary: 'Get a single option by ID'
            tags: [Options]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Option ID'}]
            responses: {'200': {description: 'Option retrieved successfully', content: {application/json: {schema: {$ref: '#/components/schemas/OptionResponse'}}}}, '400': {description: 'Invalid option ID format'}, '404': {description: 'Option not found'}, '500': {description: 'Internal server error'}}
        patch:
            summary: 'Update an option'
            tags: [Options]
            security: [{BearerAuth: []}]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Option ID'}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {option_type: {type: string, maxLength: 50}, option_value: {type: string, maxLength: 100}, name: {type: string, maxLength: 100}, is_active: {type: boolean}, sort_order: {type: number, minimum: 0}}}}}}
            responses: {'200': {description: 'Option updated successfully', content: {application/json: {schema: {$ref: '#/components/schemas/OptionResponse'}}}}, '400': {description: 'Validation error or duplicate option'}, '401': {description: 'Unauthorized - Admin access required'}, '404': {description: 'Option not found'}, '500': {description: 'Internal server error'}}
        delete:
            summary: 'Soft delete an option'
            tags: [Options]
            security: [{BearerAuth: []}]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Option ID'}]
            responses: {'204': {description: 'Option deleted successfully'}, '400': {description: 'Invalid option ID format'}, '401': {description: 'Unauthorized - Admin access required'}, '404': {description: 'Option not found'}, '500': {description: 'Internal server error'}}
...